change usize to u16

+++ rust/src/simple_state/find/root.rs
@@ -72,7 +72,7 @@ impl<const MAX_CLIQUE: u32> State<MAX_CLIQUE> {
     pub fn find_strategy_root(
         &self,
         search_state: &SearchState<MAX_CLIQUE>,
-        depth: usize,
+        depth: u16,
         max_size: u8,
     ) -> FindResult {
         let total_start = Instant::now();
@@ -138,7 +138,7 @@ fn print_headers() {
 
 enum CountMessage {
     Row {
-        depth: usize,
+        depth: u16,
         time: String,
         elapsed: String,
         result_label: String,
